Implement a basic __main__.py file able to process a --excavator arg.
This commit adds a basic argparse based package level main file which
given a filename and a named excavator will perform an excavation and
output the usual HTML files.

By default this uses the standard temporary path, but additionally
support specifying -d and allow a directory argument of "." which will
cause outputting output folder in the cwd: ./output/_autoarchaologist

In order to ensure things are working include tests of the new command
line as well as the basic operation of excavating the example file. The
tests check the expected HTML files were written for this known input
and assert some basic properties of the excavated artifacts.

While here repair the example excavation. The excavation it executes is
explicitly declared thus CLI selectable as --excavator examples.showcase